requirements.txt
pillow~=9.5.0
nextcord>=2.4.0
python-dotenv~=0.20.0
humanfriendly~=10.0
psutil~=5.9.1
emoji~=2.0.0
wavelink>=1.3.0
aiohttp~=3.10.11
asyncmy~=0.2.10
aiofiles~=24.1.0